% Please edit documentation in R/document_datasets.R
\docType{data}
\name{and_cpm_with_simulations}
\alias{and_cpm_with_simulations}
\title{Output of runnning a cross-sectional data set with evam and generating 10000 sample for each available CPM}
\format{
A list with 31 items:
\describe{
  \item{*_model}{data.frame, DAG representation of the model}
  \item{*_trans_mat}{matrix, Transition rate matrix}
  \item{*_td_trans_mat}{matrix, Time discretized transition rate matrix}
  \item{MHN_theta}{matrix, Thetas values for each gene}
  \item{HESBCN_parent_set}{list, Relationship of each genes with its parents}
  \item{name}{string, Data set name}
  \item{csd_data}{data.frame, Binary data with cross sectional data}
  \item{*genotype_transitions}{data.frame, Observed umber of transitions between two genotypes}
  \item{*sampled_genotype_freqs}{data.frame, Genotype frequencies of each genotype sampled by each CPM}
}
}
\usage{
and_cpm_with_simulations
}
\description{
A sample output of EvAM-tools with sampling for each CPM.
This file was generated by running the script evamtools/inst/shiny-examples/evamtools/test_data/generate_test_data.R
}
\keyword{datasets}
