% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary2int}
\alias{binary2int}
\title{Binary to Integer}
\usage{
binary2int(binary_state)
}
\arguments{
\item{binary_state}{vector with 0 and 1}
}
\value{
Integer values of the genotype
}
\description{
Transform a genotype from binary nomenclature to integer value
}
