% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM-plotting.R
\name{compute_vertex_labels}
\alias{compute_vertex_labels}
\title{Return the labels of most relevant paths starting from WT}
\usage{
compute_vertex_labels(
  graph,
  paths_from_graph,
  top_paths = NULL,
  type = "genotype"
)
}
\arguments{
\item{graph}{igraph object with genotype transitions}

\item{paths_from_graph}{List of path from WT to ending genotype}

\item{top_paths}{Int > 0. Include labels from vertex present in n top_paths}

\item{type}{String. Default genotype. Valid options are "genotypes" or "acquisition"
"genotype" option returns the genotype of the vertex.
"acquisition" option returns the genotype acquire along the path.}
}
\description{
Return the labels of most relevant paths starting from WT
}
