% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM-plotting.R
\name{cpm_layout}
\alias{cpm_layout}
\title{Computes the best layout for a genotype transtion matrix}
\usage{
cpm_layout(graph)
}
\arguments{
\item{graph}{igraph object with genotype transitions}
}
\value{
dataframe the x and y position for each vertex in graph
}
\description{
Genotypes are distributed in the x axis according the number 
of mutated genes they have
Distribution in the Y axis in done alphabetically.
}
