% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_datasets.R
\docType{data}
\name{cpm_output}
\alias{cpm_output}
\title{Output of CPMS}
\format{
Each element cotains 27 elements. This includes different outputs for 
each of the following CPMs: OT, CBN, MHN, HESBCN. DBN and MCCBN are being considered.
Some of the specific ouputs are:
\describe{
  \item{csd_data}{csd_data, Dataframe with the cross sectional data that generated the results}
  \item{*_trans_rate_mat}{Transition rate matrix.}
  \item{*_f_graph}{For OT (and DBN?) what we call weighted fitness graph. These are not really probabilities, though they are obtained with an algorithm identical to the one used for the transition rate matrix.}
  \item{*_model}{Contains the edges of the gene relationship DAG}
  \item{*_trans_mat}{Transition matrix between genotypes.}
  \item{*_td_trans_mat}{Time discretized transition matrix.}
  \item{OT_genots_predicted}{Genotype frequency for OT model}
}
}
\usage{
cpm_output
}
\description{
A list containing the output of several CPMs for 10 cross sectional data sets.
}
\keyword{datasets}
