% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn-process.R
\name{do_DBN}
\alias{do_DBN}
\title{Run DBN on data}
\usage{
do_DBN(data)
}
\arguments{
\item{data}{data.frame with cross sectional data}
}
\value{
list$edges Data.frame with From, To, Edge and Theta.

list$thetas Thetas of the best fit

list$likelihood Float. Likelihood of the fit
}
\description{
Run DBN on data
}
