\name{evamtools-deprecated}
\alias{evamtools-deprecated}
\title{Deprecated functions in package \sQuote{evamtools}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{evamtools} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{plot_CPMs: \code{\link{plot_evam}}}
    \item{sample_CPMs: \code{\link{sample_evam}}}
    

  }
}
