% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genotypes_from_trm.R
\name{indiv_sample_from_trm}
\alias{indiv_sample_from_trm}
\title{Sample an indivial from a transition rate matrix}
\usage{
indiv_sample_from_trm(trm, T_sampling, ngenots = NULL, genot_names = NULL)
}
\arguments{
\item{trm}{transition rate matrix}

\item{T_sampling}{Time at which sampling happens.}

\item{ngenots}{Number of genotypes}

\item{genot_names}{String array with genotype names}
}
\value{
sampled genotype, trajectory, and accumulated time
}
\description{
This is a transition *rate* matrix, *not* a matrix
of transition probabilities between genotypes. For example, this
cannot be used with OT.
}
