% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genotypes_from_trm.R
\name{indiv_sample_from_trm_pre}
\alias{indiv_sample_from_trm_pre}
\title{Sample an individual from a transition rate matrix with precomputed
matrix diagonal (and number of genotypes and genotype names).}
\usage{
indiv_sample_from_trm_pre(trmstd, diag, T_sampling, ngenots, genot_names)
}
\arguments{
\item{trmstd}{transition rate matrix "standardized",}

\item{diag}{diagonal of transition rate matrix, time of sampling of a case/individual}

\item{T_sampling}{Time at which sampling happens.}

\item{ngenots}{Number of genotypes}

\item{genot_names}{String array with genotype names}
}
\value{
sampled genotype, trajectory, and accumulated time
}
\description{
If we sample many individuals from the same process, it is worth
    passing some common, precomputed values. This is what I call
    "transition matrix standardized": Diagonal is passed separately, and
    entries in the matrix are probabilities.

This is all for a transition *rate* matrix, *not* a matrix
of transition probabilities between genotypes. For example, this
cannot be used with OT.
}
