% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int2str}
\alias{int2str}
\title{Integer to string}
\usage{
int2str(int_state, sep = ", ", wt = "WT")
}
\arguments{
\item{int_state}{Ingeter >=0}

\item{sep}{String. Separator between genes letters}

\item{wt}{String. How to define the wild type}
}
\value{
string with mutated genes
}
\description{
Transform a genotype from integer nomenclature to string binary coding
}
