% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM-plotting.R
\name{plot_sampled_genots}
\alias{plot_sampled_genots}
\title{Use plot matrix to plot the sampled genotypes}
\usage{
plot_sampled_genots(data)
}
\arguments{
\item{data}{data.frame object with cross sectional datas}
}
\description{
Use plot matrix to plot the sampled genotypes
}
