% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genotypes_from_trm.R
\name{population_sample_from_trm}
\alias{population_sample_from_trm}
\title{Sample a population from a transition rate matrix.}
\usage{
population_sample_from_trm(
  trm,
  n_samples = 10,
  T_sampling = NULL,
  pre_compute = TRUE,
  cores = detectCores()
)
}
\arguments{
\item{trm}{transition rate matrix, number of samples or times of samples,}

\item{n_samples}{Int with the number of samples to be computed}

\item{T_sampling}{Time at wich each individual in sample. By default they 
are randomly generated from an exponential distribution of rate 1.}

\item{pre_compute}{whether or not to precompute entries of the trans rate matrix (for speed)}

\item{cores}{number of cores (pass 1 is you do not want to parallelize)}
}
\value{
List with precompunted sampling time of sampling, the actual time sampled
observed for each sample (s), the complete trajectory of acquired mutations
and the observed genotype
}
\description{
Like indiv_sample_from_trm, but for multiple individuals.
}
