% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM-plotting.R
\name{process_data}
\alias{process_data}
\title{Process data of CPMs to make it easier to plot}
\usage{
process_data(data, mod, plot_type, sample_data = NULL)
}
\arguments{
\item{data}{Complete CPM output
by calling sample_CPMs with the exact CPM output from data parameter}

\item{mod}{String for the CPM to process.}

\item{plot_type}{String for the plot_type to process.}

\item{sample_data}{Data form sampling, this must be generated}
}
\value{
List with processed output of the CPM
}
\description{
Extracts the outpus concerning a single CPM
}
