% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genotypes_from_trm.R
\name{process_samples}
\alias{process_samples}
\title{Process samples}
\usage{
process_samples(
  sim,
  n_genes,
  gene_names,
  output = c("frequencies", "state_counts", "transitions")
)
}
\arguments{
\item{sim}{list generated with population_sample_from_trm. Relevant
fields are described below
$T_sum_events time of events for the mutations of each gene
$obs_events data.frame with mutated before the end of the sampling time}

\item{n_genes}{number of genes observed}

\item{gene_names}{List of gene names. Required.}

\item{output}{type of output that we want}
}
\value{
List with a list of trajectories (the order in which gene mutations
are acquired), genotype frequencies and genotypes transition matrix (with
counts of how many transitions between each genotype have been observed)
}
\description{
Generate trajectories from data simulated from a given model.
}
