% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM-plotting.R
\name{rank_paths}
\alias{rank_paths}
\title{Order paths of a weigthed directd graph}
\usage{
rank_paths(graph)
}
\arguments{
\item{graph}{igraph object with genotype transition. The graph is expect to be directed and weighted}
}
\value{
List with all paths sorted in descreasing order of importance
}
\description{
Computes the relevance of the paths starting on WT based on edge weigth
}
