\name{runs_for_refactor_main_funct}
\alias{runs_for_refactor_main_funct}
\docType{data}
\title{
Output from five runs of a former version of the main function, useful
when refactoring.
}

\description{
Output from five runs of a former version of the main function, useful
for refactoring. Code of how they were created in /misc/data_creation_for_test_main-funct.R
}
\usage{data("runs_for_refactor_main_funct")}

\format{
  A list of length five with output from five runs, named test_main_input_data, test_main_Dat1,  test_main_Dat2, test_main_Dat3, test_main_Dat4.
}

\examples{
\dontrun{
data(runs_for_refactor_main_funct)
}
}
\keyword{datasets}
