% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Run the web application of evamtools}
\usage{
runShiny()
}
\description{
Transform a genotype from integer nomenclature to binary coding Launchs the server with the web based app
}
