% Please edit documentation in R/document_datasets.R
\docType{data}
\name{sample_evam_output}
\alias{sample_evam_output}
\title{Typical data strucuture used in the shiny web app.
  It includes 1) the results of a cross-sectional data set with evam and generating 10000 sample for each available CPM,
  2) originan data infomation of the input used to generate it, and
  3) summary of CPMs results as a group of tables.}
\format{
A list with 3 items:
\describe{
  \item{cpm_output}{Merged output of runnning evam and sample_CPMs. 52 itmes.
    \itemize{
      \item{*_model}{data.frame, DAG representation of the model.}
      \item{*_trans_mat}{matrix, Probability transition matrix.}
      \item{*_trans_rate_mat}{matrix, Transition rate matrix.}
      \item{*_td_trans_mat}{matrix, Time discretized transition rate matrix.}
      \item{OT_f_graph}{sparseMatrix, Matrix with rate transitions.}
      \item{MHN_theta}{matrix, Thetas values for each gene.}
      \item{MHN_exp_theta}{matrix, Transformed thetas values for each gene.}
      \item{OncoBN_f_graph}{sparseMatrix, Matrix with rate transitions.}
      \item{OncoBN_parent_set}{vector, Relationship of each genes with its parents.}
      \item{OncoBN_likelihood}{float, Likelihood of the model.}
      \item{OncoBN_epsilon}{float, Epsilon of the model.}
      \item{OncoBN_fitted_model}{string, Model used.}
      \item{HESBCN_parent_set}{vector, Relationship of each genes with its parents.}
      \item{HESBCN_command}{list, Relationship of each genes with its parents.}
      \item{name}{string, Data set name.}
      \item{csd_data}{data.frame, Binary data with cross sectional data.}
      \item{*_predicted_genotype_freqs}{data.frame, Genotype probabilities of each genotype.}
      \item{*_sampled_genotype_counts}{vector, Genotype counts (or absolute frequencies) of each genotype sampled by each CPM.}
      \item{*_obs_genotype_transitions}{vector, Observed number of transitions between two genotypes.}
      \item{*_state_count}{data.frame, Observed number of times that a states has been visited.}
      \item{original_data}{data.frame, All cross sectional data passed evam.}
      \item{analyzed_data}{data.frame, Original cross sectional data to run evam.}
      \item{all_options}{list, Options used to run evam.}
    }
  }

  \item{orig_data}{Original data handled as input for the shiny web app.
    \itemize{
      \item{data}{data.frame, Cross sectional data used for running evam.}
      \item{name}{string, Name of the data set.}
      \item{type}{string, Type of input: either  "csd", "dag", or "matrix".}
      \item{gene_names}{vector, Genes names of the data set.}
      \item{thetas}{matrix, Matrix of thetas or NULL.}
      \item{lamdas}{vector, Lambda of each genes or NULL.}
      \item{dag}{matrix, Adjacency matrix of the dag or NULL.}
      \item{dag_parent_set}{vector, Relationship of each genes with its parents or NULL.}
    }
  }

  \item{tabular_data}{Information of CPMs and samples in a tabular way to display in the shiny app.
    \itemize{
      \item{trans_mat}{data.frame, Probabilities of transitions between two genotypes for all supported models.}
      \item{trans_rate_mat}{data.frame, Rate of transitions between two genotypes for all supported models.}
      \item{obs_genotype_transitions}{data.frame, Number of times a transition between two genotypes has been observed when sampling supported models.}
      \item{predicted_genotype_freqs}{data.frame, Probabilities of each genotype at equilibrium given a model.}
      \item{sampled_genotype_counts}{data.frame, Times a genotype has been observed when sampling.}
    }
  }
  }
}
\usage{
sample_evam_output
}
\description{
A sample output of EvAM-tools with sampling for each CPM.
This file was generated by running the script evamtools/inst/shiny-examples/evamtools/test_data/generate_test_data.R.
}
\keyword{datasets}
