% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_datasets.R
\docType{data}
\name{sampling_output}
\alias{sampling_output}
\title{Output from the sampling process}
\format{
Each element cotains 27 elements. This includes different outputs for 
each of the following CPMs: OT, CBN, MHN, HESBCN. DBN and MCCBN are being considered.
Some of the specific ouputs are:
\describe{
  \item{trajectory}{trajectory, Every nested list shows genotypes jumps seen in a single sample}
  \item{obs_events}{obs_events, List with the laste genotypes of the trajectory: the observed events}
}
}
\usage{
sampling_output
}
\description{
A list containing the output of several CPMs for 10 cross sectional data sets.
}
\keyword{datasets}
