\docType{data}
\name{SHINY_DEFAULTS}
\alias{.ev_SHINY_dflt}
\title{Defaults options for running the shiny web app}
\format{
Defaults values of the shiny app
\describe{
  \item{max_genes}{Maximun number of genes allowed}

  \item{min_genes}{Minimum number of genes allowed}
  
  \item{ngenes}{Integer of default number of genes to use when building}
  
  \item{cpm_samples}{Number of patients to samples to generate csd data from a matrix using with CPM outputs}

  \item{all_cpms}{All CPMs in evamtools}
  
  \item{csd_samples}{Number of patients to samples to generate csd data from a matrix or a dag}

  \item{cpms2run}{Which CPMs to run in the shiny server by defaults}

  \item{template_data}{One of:
    \itemize{
      \item{csd_counts:}{Data frame with the counts of each genotype}
      \item{data:}{Data frame with cross sectional data.}
      \item{dag:}{Matrix of 10x10 with lambdas}
      \item{dag_parent_sest:}{List of 10 elements with "Single"}
      \item{lambdas:}{Vector of 10 lambdas, equals to 1}
      \item{thetas:}{Matrix of 10x10 with thetas}
      \item{gene_names:}{List with gene names}
      \item{name:}{String with the data set name}
    }
  }
}}
\usage{
data(SHINY_DEFAULTS)
}
\description{
Defaults of the web app.
This file was generated by running the script
/inst/shiny-examples/evamtools/DEFAULTS.R

You will want to rerun it (so that the RData file is created again)
whenever you make changes to it.

The object is called .ev_SHINY_dflt to minimize the risk of overwriting.
}
\keyword{datasets}
