
\docType{data}
\name{SHINY_DEFAULTS}
\alias{SHINY_DEFAULTS}
\title{Defaults options for running the shiny web app}
\format{
A list of cross sectional data set to be used as example inputs.
Each data set includes 1) the data and 2) a name. In some cases there is also dag, or the values of a matrix.
A list with 3 items:
\describe{
  \item{max_genes}{Maximun number of genes allowed}

  \item{min_genes}{Minimum number of genes allowed}
  
  \item{ngenes}{Integer of default number of genes to use when building}
  
  \item{mhn_samples}{Number of patients to samples to generate csd data from a matrix using MHN functionality}

  \item{all_cpms}{All CPMs in evamtools}

  \item{cpms2run}{Which CPMs to run in the shiny server by defaults}

  \item{template_data}{One of:
    \itemize{
      \item{csd_freqs:}{Data frame with the frequencies of each genotype}
      \item{data:}{Data frame with cross sectional data.}
      \item{dag:}{Matrix of 10x10 with lambdas}
      \item{dag_parent_sest:}{List of 10 elements with "Single"}
      \item{lambdas:}{Vector of 10 lambdas, equals to 1}
      \item{thetas:}{Matrix of 10x10 with thetas}
      \item{gene_names:}{List with gene names}
      \item{name:}{String with the data set name}
    }
  }
}}
\usage{
data(examples_csd)
}
\description{
A list of cross sectional data set to be used as example inputs.
This file was generated by running the script /inst/miscell/examples/toy_datasets.R
}
\keyword{datasets}