% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str2binary}
\alias{str2binary}
\title{String to binary}
\usage{
str2binary(str_state, sep = ", ", wt = "WT", n = NULL)
}
\arguments{
\item{str_state}{String}

\item{sep}{String. Separator between genes letters}

\item{wt}{String. How to define the wild type}

\item{n}{Integer. Number of digits to return}
}
\value{
vector with 0 and 1 with the binary coding
}
\description{
Transform a genotype from string nomenclature to binary coding
}
