% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_datasets.R
\docType{data}
\name{cpm_output}
\alias{cpm_output}
\title{Output of CPMS}
\format{
Each element cotains 27 elements. This includes different outputs for 
each of the following CPMs: OT, CBN, MHN, HESBCN. DBN and MCCBN are being considered.
Some of the specific ouputs are:
\describe{
  \item{csd_data}{csd_data, Dataframe with the cross sectional data that generated the results}
  \item{*_f_graph}{Matrix with genotype-genotypes transition rates. Or probabilities for OT}
  \item{*_model}{Contains the edges of the gene relationship DAG}
  \item{*_trans_mat}{Conditional probability transition matrix}
  \item{*_td_trans_mat}{Time discretized transition matrix}
  \item{OT_genots_predicted}{Genotype frequency for OT model}
  \item{MHN_transitionRateMatrix}{Transition rate matrix of MHN}
}
}
\usage{
cpm_output
}
\description{
A list containing the output of several CPMs for 10 cross sectional data sets.
}
\keyword{datasets}
