% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hesbcn-process.R
\name{do_HESBCN}
\alias{do_HESBCN}
\title{Run HESBCN}
\usage{
do_HESBCN(
  data,
  n_steps = 1e+05,
  tmp_dir = NULL,
  seed = NULL,
  addname = NULL,
  silent = TRUE
)
}
\arguments{
\item{data}{Cross secitonal data. Matrix of genes (columns)
and individuals (rows)}

\item{n_steps}{Number of steps to run. Default: 100000}

\item{tmp_dir}{Folder name where the oput is located. 
It will be place under /tmp/HESBCN/tmp_folder}

\item{seed}{Seed to run the experiment}

\item{addname}{String to append to the temporary folder name. Default NULL}

\item{silent}{Whether to run show message showing the folder name where HESBCN is run}
}
\value{
A list with the adjacency matrix, the lambdas, the parent set
and a data.frame with From-To edges and associated lambdas.
}
\description{
Run HESBCN
}
