% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-run-all-methods.R
\name{all_methods_2_trans_mat}
\alias{all_methods_2_trans_mat}
\title{Runs the CPMs}
\usage{
all_methods_2_trans_mat(x, cores_cbn = 1, do_MCCBN = FALSE, max.cols = 15)
}
\arguments{
\item{x}{cross sectional data}

\item{cores_cbn}{How many cores to use for CBN}

\item{do_MCCBN}{Wether to run MCCBN. Default is FALSE.}

\item{max.cols}{Maximum number of columns to use in the analysis. If x has >
max.cols, selected columns are those with the largest number of events.}
}
\description{
Executes all CPMS given a cross sectional data set
}
