\docType{data}
\name{examples_csd}
\alias{examples_csd}
\title{Cross sectional data sets}
\format{
  A list of cross sectional data sets. Each data set includes 1) the data and 2) a name. In some cases there is also dag, or the values of a matrix.
A list with 3 items:
\describe{
  \item{csd}{csd, created by directly entering cross-sectional data.}
  \item{dag}{dag, from models with DAGs.}
  \item{matrix}{matrix, from MHN.}
}}
\usage{
data(examples_csd)
}
\description{
A list of cross sectional data set to be used as example inputs, for
instance by the Shiny web app.
This file was generated by running the script /inst/miscell/examples/toy_datasets.R
}
\keyword{datasets}
