% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_datasets.R
\docType{data}
\name{examples_csd}
\alias{examples_csd}
\title{Cross sectional data sets}
\format{
A list with 3 items:
\describe{
  \item{csd}{csd, with 17 different scenarios. They represents different relationships: AND, OR, and XOR, but there are also more complex set ups, like sign epitasis and reciprocal sign epistasis. }
  \item{dag}{dag, It has a simple example of an AND relationships}
  \item{matrix}{matrix, Simple random example}
}
}
\usage{
data(examples_csd)
}
\description{
A list of cross sectional data set to be used as example inputs.
This file was generated by running the script /inst/miscell/examples/toy_datasets.R
}
\keyword{datasets}
