% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_sorted_genotypes}
\alias{generate_sorted_genotypes}
\title{List of sorted genotypes}
\usage{
generate_sorted_genotypes(n_genes, gene_names = NULL)
}
\arguments{
\item{n_genes}{Ingeter >=0}

\item{gene_names}{Vector of strings
If NULL, defaults gene_names will the firt n_genes letters of the alphabet}
}
\value{
Vector with the sorted genotypes
}
\description{
Returns all sorted genotypes for a given number of genes
}
