% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccbn-process.R
\name{mccbn_hcbn_proc}
\alias{mccbn_hcbn_proc}
\title{Run MCCBN with H-CBN2

Using H-CBN2, as in example from https://github.com/cbg-ethz/MC-CBN}
\usage{
mccbn_hcbn_proc(
  x,
  max.iter.asa = 10000,
  ncores = 1,
  L = 100,
  tmp_dir = NULL,
  addname = NULL,
  silent = TRUE
)
}
\arguments{
\item{max.iter.asa}{Int. Number of steps to run. Default: 100000}

\item{ncores}{Int. Number of threads to use}

\item{L}{Int. Number of samples to be drawn from the proposal in the E-step}

\item{tmp_dir}{Directory name where the oput is located.}

\item{addname}{String to append to the temporary folder name. Default NULL}

\item{silent}{Whether to run show message showing the folder name where MCCBN is run}

\item{data}{Cross secitonal data. Matrix of genes (columns)
and individuals (rows)}
}
\value{
A list with the adjacency matrix, the lambdas, the parent set
and a data.frame with From-To edges and associated lambdas.
}
\description{
Run MCCBN with H-CBN2

Using H-CBN2, as in example from https://github.com/cbg-ethz/MC-CBN
}
