% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPM-plotting.R
\name{process_data}
\alias{process_data}
\title{Process data of CPMs to make it easier to plot}
\usage{
process_data(data, mod, prune_edges = TRUE)
}
\arguments{
\item{data}{Complete CPM output}

\item{mod}{String for the CPM to process.}

\item{prune_edges}{Boolean. Wether to remove genotype relationships carrying less than 1% of the flux}
}
\value{
List with processed output of the CPM
}
\description{
Extracts the outpus concerning a single CPM
}
