\name{runShiny}
\alias{runShiny}
\title{Run the web application of evamtools}
\usage{
runShiny(host="0.0.0.0", port=3000, test.mode = FALSE)
}
\arguments{
\item{host}{Host where the app will be listening}
\item{port}{Port where the app will be listening}
\item{test.mode}{See \code{\link[shiny]{runApp}}}
}
\description{
Launch the server with the web based app.
}
