% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_genotypes_from_trm.R
\name{sample_all_CPMs}
\alias{sample_all_CPMs}
\title{Run samples for all outputs of CPMs

OT already provides the predicted genotypes
though not from a transition rate matrix}
\usage{
sample_all_CPMs(
  cpm_output,
  n_samples,
  n_genes,
  gene_names = NULL,
  methods = c("Source", "CBN", "MCCBN", "DBN", "MHN", "HESBCN", "OT")
)
}
\arguments{
\item{cpm_output}{Output from calling all_methods2trans_mat}

\item{n_samples}{Number of samples to generate}

\item{n_genes}{Number of samples that are in the sample}

\item{gene_names}{List of gene names. If NULL, genes will be named alphabetically}

\item{methods}{List of methods that we want to sample}
}
\value{
modified cpm_outputd including a matrix with genotype transitions
}
\description{
Run samples for all outputs of CPMs

OT already provides the predicted genotypes
though not from a transition rate matrix
}
